backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
PyJWT==2.7.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-json-logger==2.0.7
